+++ Assets/Scripts/Unit/Unit.cs
@@ -1,4 +1,14 @@
 using UnityEngine;
 public class Unit : Health
 {
+    public float maxHealth = 100f;
+    private float currentHealth;
+
+    public HealthBar healthBar; // HealthBar 스크립트 연결
+
+    void Start()
+    {
+        currentHealth = maxHealth;
+        healthBar.UpdateHealth(currentHealth / maxHealth);
+    }
 }
